marker-layer: Properly position markers
Commit 9e138d2 reworked allocation of ShumateMarkerLayer, but it wrongly
used the x/y positions relative to the *marker* and offset by half of the
width/height. The correct coordinate transform is relative to the mark layer,
then offsetting by half of the marker size to centralize it.

Use the correct relative widget to translate between geographic coordinates
and pixel coordinates.