view: Reimplement animated go-to using tick callbacks
For now, it uses a hardcoded ease-in-out-quad easing function, which
was the default in the past. To use tick callbacks, we have to track
the start (µs), and the duration (ms, converted to µs).

The progress is calculated by checking the current frame time (µs).
Easing is then applied to the progress itself, which is then used
in a simple LERP operation to get the next latitude/longitude.

Fixes https://gitlab.gnome.org/GNOME/libshumate/-/issues/10