view: Add kinetic scrolling
Implement kinetic scrolling through a GtkGestureSwipe that is triggered
after ending a drag. When triggered, it adds a tick callback to the view,
and each tick calculates the new distance from the point where the drag
end happened.

The kinetic scrolling heuristic is copied and modified from GTK4.

The deceleration rate is hardcoded for now, because the 'deceleration'
property is stub, but that'll be fixed by the next commits.

Fixes: https://gitlab.gnome.org/GNOME/libshumate/-/issues/7