map-source: Make the fill_tile vfunc async
Previously, fill_tile_async called the original fill_tile virtual
method, which did not call a callback when it was finished. The async
callback was implemented by connecting to the tile's notify::state
signal.

Now, the virtual method itself is async, not just the wrapper function.
This required some changes to the functions, most notably in
ShumateNetworkTileSource.