map-layer: Move logic out of size_allocate
Move a lot of the tile grid logic out of size_allocate into
recompute_grid(). This is neater since size_allocate isn't really
supposed to modify the content of the widget, only position it on
screen.

Part of this change is that tiles are stored in the hash map by their
unwrapped grid coordinates, not by coordinates in the screen grid. This
makes much of the logic simpler, especially the logic around
invalidating tiles. Also, tiles are no longer reused; once a tile is
created, its x/y/zoom will not change, and it will be deleted and
replaced when it moves out of view.