map-layer: Fix flashing when zoom level changes
Previously, the screen would flash every time the zoom level changed.
This was because the tiles from the previous zoom level were all deleted
as soon as the zoom level changed.

Now, those old tiles are kept around until all the tiles at the new zoom
level are loaded. This provides something to fill the screen while the
tiles load.

This commit also simplifies the size_allocate logic. Instead of
iterating through the expected tile locations and finding them in the
hash table, it iterates the hash table directly and calculates the
correct location for each tile.