map-layer: Reduce tile requests during animations
When the map layer detects that the camera is moving quickly (during an
animation, or for some other reason), it will skip requesting tiles on
that frame. This helps prevent tiles from being requested if they will
likely be discarded before they load.

A recalculate_grid() call is scheduled for the next frame, so if the
motion suddenly stops, the tiles will still be loaded.