vector: Optimize common expression combinations
Previously, according to Sysprof, filter evaluation took about as long
as actually drawing the tiles (when drawing my in-progress map style,
which has a lot of layers).

Improved performance by creating new expression types for common
combinations of expressions, such as checking whether the value of a
certain tag is in a certain set. The combinations that are optimized
cover most of what map styles use in practice.

On the in-progress map style, this change brings a 2-3x speedup in
filter evaluation.