vector: Fix shumate_vector_collision_new signature
According to the C standard, declaring a function without an argument
list does _not_ mean that the function expects zero arguments. Rather,
it means that the passed on arguments shouldn't be checked.

In case you want to specify a function without arguments, you should use
`void` instead. Let's use that here.

This fixes a build failure with GCC 15 that was happening in Fedora
Rawhide.