vector: Make the vector renderer not optional
This commit removes the conditions for building the vector renderer.
It is always built now. The meson build option remains for backward
compatibility, but it does nothing except emit a warning when it is
false.

Originally, the vector renderer was an experimental addon that could
be enabled with a build flag. When Maps started using it, the flag was
changed to be enabled by default. Now there is little use for building
libshumate without it, and the conditional code just adds to the
complexity of the build and APIs.